[tests] update expected output of rasterization.plyflatten_from_plyfiles_list
The definition of the DSM sampling grid computed by this function has
been updated so that its vertices are integer multiples of the grid cell
size. Hence the output DSM is slightly different and the expected test
result must be updated accordingly. In particular, as the output GeoTIFF
corner coordinates are rounded according to the GSD parameter, their
bounds can differ up to one GSD and the raster size multiplied by the
GSD can differ up to 2 GSD.